feat: controller scanner 객체 생성
